changelog-builder/pyproject.toml
[project]
name = "changelog-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hyperbrowser",
    "jsonref>=1.1.0",
    "openai>=1.65.4",
    "playwright>=1.50.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.1",
    "tenacity>=9.0.0",
    "typing-extensions>=4.12.2",
]

[tool.uv.sources]
hyperbrowser = { path = "../../code/sdks/python-sdk/dist/hyperbrowser-0.35.0-py3-none-any.whl" }
